made rotation interval long from int for consistency

+++ src/main/java/io/reign/metrics/StaticMetricRegistryManager.java
@@ -77,8 +77,8 @@ public class StaticMetricRegistryManager implements MetricRegistryManager {
     }
 
     @Override
-    public int getRotationInterval() {
-        return Integer.MAX_VALUE;
+    public long getRotationInterval() {
+        return Long.MAX_VALUE;
     }
 
     @Override
